+++ packages/zowe-explorer-api/src/profiles/ZoweExplorerZosmfApi.ts
@@ -306,6 +306,10 @@ export class ZosmfMvsApi extends ZosmfApiCommon implements ZoweExplorerApi.IMvs 
  */
 // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 export class ZosmfJesApi extends ZosmfApiCommon implements ZoweExplorerApi.IJes {
+    public async getJobsByParameters(params: zowe.IGetJobsParms): Promise<zowe.IJob[]> {
+        return await zowe.GetJobs.getJobsByParameters(this.getSession(), params);
+    }
+
     public async getJobsByOwnerAndPrefix(owner: string, prefix: string): Promise<zowe.IJob[]> {
         return await zowe.GetJobs.getJobsByOwnerAndPrefix(this.getSession(), owner, prefix);
     }