+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 
 ## TBD Release
 
+### New features and enhancements
+
+- New API call `getJobsByParameters` to allow filtering jobs by status.
+
 ### Bug fixes
 
 - Added logging in places where errors were being caught and ignored.
 
-## `2.4.1`
+## 2.4.1
 
 ### Bug fixes